requirements.txt
python-dotenv
python-docx
alive-progress